Rename plugins to apps
Fixes #632
