+++ src/apps/stats.ts
